MemoryLeaks == NULL

+++ trunk/Aeona Sprint 2/Aeona Project/source/AI_States/CBossAirAI.cpp
@@ -196,8 +196,8 @@ void CBossAirAI::Update(CBaseCharacter* pCharacter, float fElapsedTime)
 				Tornado->Activate();
 				CObjectManager::GetInstance()->AddObject(Tornado);
 				PointD temp = pCharacter->GetPosition();
-				float temp_y_dist = pCharacter->GetPosY() - CPlayer::GetInstance()->GetPosY();
-				float temp_x_dist = pCharacter->GetPosX() - CPlayer::GetInstance()->GetPosX();
+				double temp_y_dist = pCharacter->GetPosY() - CPlayer::GetInstance()->GetPosY();
+				double temp_x_dist = pCharacter->GetPosX() - CPlayer::GetInstance()->GetPosX();
 
 				if(fabs(temp_y_dist) > fabs(temp_x_dist))
 				{
